HDWallet creation thread-safe (#1528)
* HDWallet creation thread-safe

* Add comment on buffer check

* Add max length define to Bep39.h, use it from there.

* Add multi-threaded creation test

* Re-enable multi-thread test

* Adapt iOS test to catch the error being tested.

* Extend test

* Test fix wallet2

+++ src/Mnemonic.h
@@ -13,9 +13,9 @@ namespace TW {
 /// BIP39 Mnemonic Sentence handling.
 class Mnemonic {
 public:
-    static constexpr size_t MaxWords = 24;
-    static constexpr size_t MinWords = 12;
-    static constexpr size_t BitsPerWord = 11; // each word encodes this many bits (there are 2^11=2048 different words)
+    static constexpr int MaxWords = 24;
+    static constexpr int MinWords = 12;
+    static constexpr int BitsPerWord = 11; // each word encodes this many bits (there are 2^11=2048 different words)
 
 public:
     /// Determines whether a mnemonic phrase is valid.
